+++ src/Jigoshop/Entity/Product/Virtual.php
@@ -168,9 +168,6 @@ class Virtual extends Product implements Purchasable, Shippable, Saleable
 	{
 		parent::restoreState($state);
 
-		if (isset($state['price'])) {
-			$this->price = (float)$state['price'];
-		}
 		if (isset($state['regular_price'])) {
 			$this->regularPrice = $state['regular_price'] !== '' ? (float)$state['regular_price'] : '';
 		}